[refactor] controller 폴더 분리
